[Deferred Shading/Render while resizing] Keep app rendering when resizing window
