+++ PointMassSpring_Cloth.vcxproj
@@ -159,6 +159,7 @@
     <ClCompile Include="External Libraries\imgui\imgui_tables.cpp" />
     <ClCompile Include="External Libraries\imgui\imgui_widgets.cpp" />
     <ClCompile Include="src\Cloth.cpp" />
+    <ClCompile Include="src\ClothSimulationApp.cpp" />
     <ClCompile Include="src\Floor.cpp" />
     <ClCompile Include="src\PointMass.cpp" />
     <ClCompile Include="src\Constant.cpp" />
@@ -185,6 +186,7 @@
     <ClInclude Include="External Libraries\imgui\imstb_rectpack.h" />
     <ClInclude Include="External Libraries\imgui\imstb_textedit.h" />
     <ClInclude Include="External Libraries\imgui\imstb_truetype.h" />
+    <ClInclude Include="src\ClothSimulationApp.h" />
     <ClInclude Include="src\FPSCounter.h" />
     <ClInclude Include="src\Cloth.h" />
     <ClInclude Include="src\Constant.h" />
