+++ PointMassSpring_Cloth.vcxproj.filters
@@ -90,6 +90,9 @@
     <ClCompile Include="src\Renderer\MeshUtility.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="src\ClothSimulationApp.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="Shader\objectFragment.shader">
@@ -116,6 +119,8 @@
     <None Include="Shader\deferredVertex.shader">
       <Filter>Resource Files\Shaders</Filter>
     </None>
+    <None Include="Shader\lightBoxFragment.shader" />
+    <None Include="Shader\lightBoxVertex.shader" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="External Libraries\stb\stb_image.h">
@@ -197,6 +202,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="src\Renderer\SkyBox.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\ClothSimulationApp.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
